use float64 for VWAT, added WebSocket command timeou

+++ internal/entity/deal.go
@@ -2,6 +2,6 @@ package entity
 
 // Deal – amount and price of trading pair
 type Deal struct {
-	Volume float32
-	Price  float32
+	Volume float64
+	Price  float64
 }
